requirements.txt
python-telegram-bot==20.4
motor==3.3.0
pymongo==4.5.0
dnspython==2.4.2